Add files via upload
Participated in the Microsoft Air Quality Challenge Held in February 2022. 

My second page of the Microsoft air quality challenge.
We see that Particulate Matter Concentration begins to increase when Temperature drops below a certain level.
Also when the chart shows that Particulate Matter concentration is lowest towards mid hours of the day(between 12-16th hour). This can be said to be as a result of increase in temperature towards noon.

#GitHub4Me
#PowerBi
#MicrosoftChallenge